[feat#4] 카카오 로그인 및 온보딩 구현

+++ src/main/java/carbonneutral/academy/common/config/AppConfig.java
@@ -39,7 +39,7 @@ public class AppConfig {
   }
 
   @Bean
-  public AuditorAware<Long> auditorAware() {
+  public AuditorAware<Integer> auditorAware() {
     return new ApplicationAuditAware();
   }
